initial outline
if there's one thing I'm bad at, it's singing. if there's another thing
I'm bad at, it's self-promotion. This resume is going to take several
iterations but and if I don't get everything out there then I'll never
get started. Is this clean? no. Does is read well? no. Is it complete?
no.

but it's a place to start.